moved prescription methods and publication to server subfolder so authentication methods are hidden from client
