+++ RxOn/server/main.js
@@ -1,3 +1,5 @@
 import '../imports/collections/prescriptions.js';
+import '../imports/collections/server/prescription_methods.js';
+import '../imports/collections/server/prescription_publications.js';
 import '../imports/startup/server/accounts-config.js'
 // this file is the main server-side entry point
